Allow overwrite of group and project in metadata
Added optional arguments for group and project. If specified, group_name
and project_name will be added to the medatada dict and written into
metadata.json in the tar, which is used by bootstrap.py to sort the data
into the proper place in the DB.

Also added optional verbose flag (-v) and changed behavior to be much
less verbose by default.